requirements.txt
PyGithub
toml
termcolor